ROX-21710: Simplify structure of clusters page
